+++ opentelemetry-otlp/examples/basic-otlp-http/src/main.rs
@@ -9,7 +9,7 @@ use opentelemetry_appender_tracing::layer::OpenTelemetryTracingBridge;
 use opentelemetry_otlp::WithExportConfig;
 use opentelemetry_otlp::{LogExporter, MetricExporter, Protocol, SpanExporter};
 use opentelemetry_sdk::{
-    logs::LoggerProvider,
+    logs::SdkLoggerProvider,
     metrics::{MetricError, SdkMeterProvider},
     trace::{self as sdktrace, TracerProvider},
 };
@@ -28,14 +28,14 @@ static RESOURCE: Lazy<Resource> = Lazy::new(|| {
         .build()
 });
 
-fn init_logs() -> Result<sdklogs::LoggerProvider, opentelemetry_sdk::logs::LogError> {
+fn init_logs() -> Result<sdklogs::SdkLoggerProvider, opentelemetry_sdk::logs::LogError> {
     let exporter = LogExporter::builder()
         .with_http()
         .with_endpoint("http://localhost:4318/v1/logs")
         .with_protocol(Protocol::HttpBinary)
         .build()?;
 
-    Ok(LoggerProvider::builder()
+    Ok(SdkLoggerProvider::builder()
         .with_batch_exporter(exporter)
         .with_resource(RESOURCE.clone())
         .build())